+++ src/main/typescript/migrations/world.ts
@@ -53,8 +53,8 @@ export async function migrateWorld(): Promise<void> {
     `Migrating this world to ${currentVersion}. Last migration was ${lastMigrVersion}`
   );
 
-  await migrateActors();
-  await migrateItems();
+  await migrateActors(currentVersion);
+  await migrateItems(currentVersion);
 
   setMigrationCurrentVersion();